rename package s/utils/cmdutil/
